[SBL-56] 질문의 구조를 더 체계적으로 리팩토링
