+++ src/main/kotlin/com/sbl/sulmun2yong/survey/domain/question/Question.kt
@@ -2,6 +2,7 @@ package com.sbl.sulmun2yong.survey.domain.question
 
 import java.util.UUID
 
+/** 설문의 질문 */
 interface Question {
     val id: UUID
     val questionType: QuestionType
@@ -10,7 +11,9 @@ interface Question {
     val isRequired: Boolean
     val choices: Choices?
 
+    /** 질문에 유효한 응답인지 검증 */
     fun isValidResponse(questionResponse: QuestionResponse): Boolean
 
-    fun canBeKeyQuestion(): Boolean
+    /** 선택지 라우팅의 keyQuestion이 될 수 있는지 판단, KeyableQuestion을 제외하고 전부 false */
+    fun canBeKeyQuestion() = false
 }
